finished changed my components on contex
